% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_hemis.R
\name{countries_hemis}
\alias{countries_hemis}
\title{Determine countries to include in app's user interface}
\usage{
countries_hemis(country)
}
\arguments{
\item{country}{A tibble filtered from \code{join_data}. see \code{?join_data}}
}
\value{
A drop down menu (i.e. \code{Shiny}'s \code{selectInput}) of selected countries
}
\description{
Takes a list of country from inbuilt countries dataset (\code{Northern_hemis}, \code{Souther_hemis} & \code{Tropics}).
}
\examples{
library(covidseasonality)
library(dplyr)
Tropics \%>\%
    filter(country == "Singapore") \%>\%
    countries_hemis()
}
