% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_list.R
\docType{data}
\name{country_list}
\alias{country_list}
\title{List of countries the wrold with its corresponding Hemisphere}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 203 rows and 2 columns.
}
\usage{
country_list
}
\description{
This data set contains 203 countries classified in either Northern(N), Southern Hemisphere(S) or Tropics(T).

#' @source \href{https://github.com/tidyverse/ggplot2}{ggplot2 map data} \cr
\href{https://github.com/joachim-gassen/tidycovid19}{tidycovid19 package} \cr
\href{https://worldpopulationreview.com/country-rankings/tropical-countries}{worldpopulationreview} \cr
}
\keyword{datasets}
