% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_data.R
\docType{data}
\name{covid_data}
\alias{covid_data}
\title{Coronavirus dataset with daily cases}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 54423 rows and 8 columns.
}
\source{
\href{https://github.com/joachim-gassen/tidycovid19}{tidycovid19 package}
}
\usage{
covid_data
}
\description{
Dataset contains daily cases of 203 countries
}
\keyword{datasets}
