% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_data.R
\name{join_data}
\alias{join_data}
\title{Join covid and map data for covidseasonality App}
\usage{
join_data(countries)
}
\arguments{
\item{countries}{countries included in datasets \code{Northern_hemis}, \code{Southern_hemis} & \code{Tropics}}
}
\value{
A tibble containing map coordinates and dates
}
\description{
Join covid and map data for covidseasonality App
}
\examples{
join_data()

}
