% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch \code{covidseasonality} app
#'}
\usage{
launch_app()
}
\value{
A shiny application
}
\description{
Launch \code{covidseasonality} app
#'
}
\examples{
launch_app()

}
