% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_countries_hemis.R
\name{plot_countries_hemis}
\alias{plot_countries_hemis}
\title{Line plot of cases against date with shaded area representing winter}
\usage{
plot_countries_hemis(x)
}
\arguments{
\item{x}{Data set containing a \code{country} column}
}
\value{
a \code{ggplot} object
}
\description{
Line plot of cases against date with shaded area representing winter
}
