% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_table}
\alias{summary_table}
\title{Provides a summary table of cases}
\usage{
summary_table(x)
}
\arguments{
\item{x}{See \code{?join_data} with dates and daily coronavirus cases.}
}
\value{
A summarized tibble of cases and cases per 100,000 of population.
}
\description{
Provides a summary table of cases
}
