% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tropics.R
\docType{data}
\name{CountriesHemisphere}
\alias{CountriesHemisphere}
\alias{Tropics}
\alias{Northern_hemis,Southern_hemis,Tropics}
\title{Countries in the Tropics and each Hemisphere}
\format{
3 data frames with 1 variable(\code{country}) each:
\describe{
\item{\code{Northern_hemis}}{countries in Northern hemisphere}
\item{\code{Southern_hemis}}{countries in Southern hemisphere}
\item{\code{Tropics}}{countries in the tropics}
}
}
\source{
\href{https://github.com/tidyverse/ggplot2}{ggplot2 map data} \cr
\href{https://github.com/joachim-gassen/tidycovid19}{tidycovid19 package} \cr
\href{https://worldpopulationreview.com/country-rankings/tropical-countries}{worldpopulationreview} \cr
}
\usage{
Northern_hemis, Southern_hemis, Tropics
}
\description{
These data sets (tibbles) contain 203 countries classified in Northern & Southern Hemisphere and Tropics.
}
\keyword{datasets}
