% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world.R
\docType{data}
\name{world}
\alias{world}
\title{World dataset including Hemisphere}
\format{
An object of class \code{data.frame} with 99338 rows and 5 columns.
}
\source{
\href{https://github.com/tidyverse/ggplot2}{ggplot2 map data} \cr
\href{https://worldpopulationreview.com/country-rankings/tropical-countries}{worldpopulationreview}
}
\usage{
world
}
\description{
This dataset contains coordinates of 252 countries and which hemisphere they are located.
}
\keyword{datasets}
