% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch \code{covidseasonality} app}
\usage{
launch_app()
}
\value{
A shiny application
}
\description{
Launch covidseasonality Shiny App embedded in the \code{covidseasonality} downloadable from \href{https://github.com/etc5523-2020/r-package-assessment-brenwin1}{GitHub}
}
